+++ src/main/scala/org/clulab/reach/ReachConstants.scala
@@ -1,28 +1,9 @@
 package org.clulab.reach
 
 /**
-  * This object wraps the taxonomy
+  * Convenience sets built from taxonomy
   */
-import java.util.Collection
-import org.yaml.snakeyaml.Yaml
-import org.yaml.snakeyaml.constructor.Constructor
-import org.clulab.odin.impl.Taxonomy
-
-
 object ReachConstants {
-
-  // Could be used by other components
-  val taxonomy = readTaxonomy("org/clulab/reach/biogrammar/taxonomy.yml")
-
-  private def readTaxonomy(path: String): Taxonomy = {
-    val url = getClass.getClassLoader.getResource(path)
-    val source = if (url == null) io.Source.fromFile(path) else io.Source.fromURL(url)
-    val input = source.mkString
-    source.close()
-    val yaml = new Yaml(new Constructor(classOf[Collection[Any]]))
-    val data = yaml.load(input).asInstanceOf[Collection[Any]]
-    Taxonomy(data)
-  }
 
   val ACTIVATION_EVENTS = taxonomy.hyponymsFor("ActivationEvent").toSet
   val REGULATION_EVENTS = taxonomy.hyponymsFor("Regulation").toSet
